fix php8 issues

+++ classes/class-cb-safe-user-deletion.php
@@ -25,17 +25,20 @@ class CB_Safe_User_Deletion {
   function show_user_delete_error_message() {
 
     if ( !session_id() ) {
+      @ob_start();
       session_start();
     }
 
-    if(array_key_exists( 'user_delete_error', $_SESSION )) {
+    if(isset($_SESSION) && array_key_exists( 'user_delete_error', $_SESSION )) {
       $class = 'notice notice-error';
       $message = $_SESSION['user_delete_error'];
 
       printf( '<div class="%1$s"><p>%2$s</p></div>', esc_attr( $class ), esc_html( $message ) );
     }
 
-    unset( $_SESSION['user_delete_error'] );
+    if(isset($_SESSION)) {
+      unset( $_SESSION['user_delete_error'] );
+    }
   }
 
   /**
